+++ packages/orq-rc/RELEASES.md
@@ -3309,3 +3309,13 @@ Based on:
 - [python v3.14.0-rc.26] packages/orq-rc
 ### Releases
 - [PyPI v3.14.0-rc.26] https://pypi.org/project/orq-ai-sdk/3.14.0-rc.26 - packages/orq-rc
+
+## 2025-10-15 08:03:52
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.636.3 (2.723.11) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.14.0-rc.27] packages/orq-rc
+### Releases
+- [PyPI v3.14.0-rc.27] https://pypi.org/project/orq-ai-sdk/3.14.0-rc.27 - packages/orq-rc